InnotterDjango/pyproject.toml
[tool.poetry]
name = "innotterdjango"
version = "0.1.0"
description = "Innowise Trainee task"
authors = ["Ilya Savin <ilya.savin@innowise-group.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.20.0"
PyJWT = "^2.4.0"
Django = "^4.1"
djangorestframework = "^3.13.1"
drf-extensions = "^0.7.1"
django-filter = "^22.1"
model-bakery = "^1.6.0"
pytest = "^7.1.2"
pytest-django = "^4.5.2"
pytest-cov = "^3.0.0"
pytest-mock = "^3.8.2"
psycopg2-binary = "^2.9.3"
Markdown = "^3.3.7"
flake8 = "^5.0.4"
dj-database-url = "^1.0.0"
celery = "^5.2.7"
boto3 = "^1.24.80"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
